requirements.txt
django==3.2.1
djangorestframework==3.14.0
easyocr==1.6.2
pytube==15.0.0
youtube-dl==2021.12.17
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
joblib==1.2.0
pandas==2.0.1
mysqlclient
pymysql
